+++ packages/react/src/components/json-tree-view/json-tree-view-node.tsx
@@ -2,7 +2,6 @@ import {
   type JsonNode,
   type JsonNodeHastElement,
   getAccessibleDescription,
-  isRootKeyPath,
   jsonNodeToElement,
   keyPathToId,
   keyPathToKey,